Members: add startdatelab field
We use the startdate as an extra selection criteria for showing
publications from the Kern lab. If people change their position,
for example by staying a little longer as a postdoc after finishing
their PhD, changing the startdate for the new position has as an
unwanted side effect that their Kern-lab publications dissapear.

Introduce 'startdatelab' to reflect the initial data of joining the lab
(only used when different than 'startdate'); 'startdate' is used to
indicate when someone started in their current position.